% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Reads in a file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the filename to be read in}
}
\value{
This function returns a dataframe containing the data in the filename if it existst. In case filename
doesn't exist, an error is displayed.
}
\description{
Reads in the specified filename in df using readr. Messages are suppressed.
If the filame doesn not exist, a message is displayed.
}
\examples{
\dontrun{
fars_read(accident_2013.csv.bz2)
}
}
