% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Makes filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A number or chracter string with numbers}
}
\value{
This function returns a file name consisting of accident_number.csv.bz2
}
\description{
Given a year number or string, this function will make a file name consisting of accident_number.csv.bz2
}
\examples{
\dontrun{
make_filename("2015")
make_filename(2015)
}
}
